+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ### Release 0.11
 
 * Changed Imperial Wind Unit from mi/h to mph
+* Fixed Rain value not displaying correctly when Imperial units
+* Changed Forecast for Weather and Sensor to start with Current Day.